5c: Add test for isPublishable
Using a data provider, multiple arguments can be passed into the test,
such as different combinations of created dates and expected results.

This test also requires mocking the `Time` object as it’s now a
dependency of the article wrapper. The `getCreatedTime` method on the
article also needs to be set to return a specific value.